src/python-api/modernized/requirements.txt
flask>=3.0
sqlalchemy>=2.0
pydantic>=2.0
pydantic-settings>=2.0
pydantic[email]
bcrypt>=4.0
PyJWT>=2.8
flask-limiter>=3.5
werkzeug>=3.0
python-dotenv>=1.0
